add viewfs

+++ dmlc-core/yarn/src/org/apache/hadoop/yarn/dmlc/Client.java
@@ -135,8 +135,13 @@ public class Client {
             LocalResource r = Records.newRecord(LocalResource.class);
             Path path = new Path(e.getValue());
             // copy local data to temporary folder in HDFS
-            if (!e.getValue().startsWith("hdfs://")) {
-                Path dst = new Path("hdfs://" + tmpPath + "/"+  path.getName());
+            if (!e.getValue().startsWith("hdfs://") && !e.getValue().startsWith("viewfs://")) {
+				// chenlongzhen
+				LOG.info("if you use hdfs or view fs data.you must start with hdfs:// or viewfs://");
+				String fsDefaultName = conf.get(FileSystem.FS_DEFAULT_NAME_KEY);
+				Path dst = new Path(fsDefaultName + tmpPath + "/"+ path.getName());
+				//end
+                //Path dst = new Path("hdfs://" + tmpPath + "/"+  path.getName());
                 dfs.copyFromLocalFile(false, true, path, dst);
                 dfs.setPermission(dst, permTemp);
                 dfs.deleteOnExit(dst);
